Warning fixes.
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@38330 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/gtk1/choice.cpp
@@ -133,7 +133,7 @@ bool wxChoice::Create( wxWindow *parent, wxWindowID id,
 
     GtkWidget *menu = gtk_menu_new();
 
-    for (unsigned int i = 0; i < n; i++)
+    for (unsigned int i = 0; i < (unsigned int)n; i++)
     {
         GtkAddHelper(menu, i, choices[i]);
     }
